+++ config/routes.rb
@@ -728,6 +728,7 @@ Rails.application.routes.draw do
       get "/site_details" => "site_details#index"
       get "/select_values" => "select_values#index"
       get "/theme" => "theme#index"
+      get "/testimonials" => "testimonials#index"
       post "/auth/firebase" => "auth#firebase"
 
       # Embeddable Widget API